Move constants.ts to shared
