+++ src/shared-react/hooks.ts
@@ -1,7 +1,7 @@
 import { isEqual, noop } from 'lodash';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { TFunction, useTranslation } from 'react-i18next';
-import { i18nConfig } from '../constants';
+import { i18nConfig } from '../shared/constants';
 
 export type TFunc = TFunction<typeof i18nConfig.ns, string>;
